More Android related fixes
